backend/pyproject.toml
[project]
name = "mankosathi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
